IntelligentQuiz/requirements.txt
django==5.2.7
# Pillow 11+ provides prebuilt wheels for modern Python versions including 3.13
pillow>=11.0.0,<12.0.0
openai==1.3.0
anthropic==0.5.0
python-dotenv==1.0.0
google-generativeai==0.7.2
